requirements.txt
fastapi
uvicorn
python-dotenv
requests
git+https://github.com/EvenTITO/eventitolibs.git
